app/requirements.txt
hug==2.5.4
gunicorn==19.9.0
slackclient>=2.0.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
google-auth==1.6.3
idna==2.8
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.21.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.3
flask==1.0.2
